backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-generativeai==0.8.3
diskcache==5.6.3
httpx==0.28.1
